adjustments

+++ backend/routes/labelRoutes.js
@@ -2,12 +2,48 @@ const express = require("express");
 const router = express.Router();
 const Label = require("../models/labelModel");
 const User = require("../models/userModel");
+let buttonIsClicked = false;
+
+// Trigger route for CPEE
+router.post("/test1", async (req, res) => {
+  try {
+    // Set the buttonIsClicked variable to true
+    buttonIsClicked = true;
+    // Send a response indicating successful recording of button click
+    res.status(200).send("User clicked label save");
+  } catch (error) {
+    console.error("Error recording label save:", error);
+    res.status(500).json({ error: "Internal server error" });
+  }
+});
+
+// Trigger route for CPEE
+router.post("/triggerTest1", async (req, res) => {
+  // Variable to store the button toggle status
+  let buttonClickToggle;
+
+  // Check if the  button is clicked
+  if (buttonIsClicked) {
+    // If clicked, set toggle to true
+    buttonClickToggle = true;
+    try {
+      res.status(200).send(buttonClickToggle);
+    } catch (error) {
+      console.error("Error saving label:", error);
+      res.status(500).json({ error: "Internal server error" });
+    }
+  } else {
+    // If not clicked, set toggle to false
+    buttonClickToggle = false;
+    res.status(200).send(buttonClickToggle);
+  }
+});
 
 // Save a new label
 router.post("/save", async (req, res) => {
   try {
     // Extract label details from request body
-    const {
+    let {
       userId,
       name,
       text,
@@ -24,6 +60,25 @@ router.post("/save", async (req, res) => {
       createdAt,
     } = req.body;
 
+    // Check if userId is undefined
+    if (userId === undefined) {
+      // Assign default values for each attribute
+      userId = "65ad315059baf730b299befe"; // Assign a default value for userId
+      name = "default labek"; // Assign a default value for name
+      // Assign default values for other attributes
+      text = "";
+      fontStyle = "Arial";
+      fontSize = 30;
+      isBold = false;
+      isItalic = false;
+      isUnderline = false;
+      textAlignment = "center";
+      verticalAlignment = "middle";
+      isQRCodeUsed = false;
+      url = "";
+      shortenedUrl = "";
+      createdAt = Date.now();
+    }
     // Create a new label instance
     const label = new Label({
       user: userId,
@@ -46,9 +101,12 @@ router.post("/save", async (req, res) => {
     await label.save();
 
     // Add the label to the user's labels array
-    await User.findByIdAndUpdate(userId, { $push: { labels: label._id } });
-
-    res.status(201).json({ message: "Label saved successfully" });
+    if (userId) {
+      await User.findByIdAndUpdate(userId, { $push: { labels: label._id } });
+    }
+    //    res.status(201).json({ message: "Label saved successfully" });
+    let isSaved = true;
+    res.status(200).send(isSaved);
   } catch (error) {
     console.error("Error saving label:", error);
     res.status(500).json({ error: "Internal Server Error" });
@@ -85,17 +143,20 @@ router.get("/user/:userId", async (req, res) => {
 router.delete("/:labelId", async (req, res) => {
   try {
     const labelId = req.params.labelId;
+    if (labelId === undefined) {
+      res.status(200).json({ message: "No label to delete" });
+    } else {
+      // Remove the label from the user's labels array
+      await User.findOneAndUpdate(
+        { labels: labelId },
+        { $pull: { labels: labelId } }
+      );
 
-    // Remove the label from the user's labels array
-    await User.findOneAndUpdate(
-      { labels: labelId },
-      { $pull: { labels: labelId } }
-    );
+      // Remove the label from the labels collection
+      await Label.findByIdAndDelete(labelId);
 
-    // Remove the label from the labels collection
-    await Label.findByIdAndDelete(labelId);
-
-    res.status(200).json({ message: "Label deleted successfully" });
+      res.status(200).json({ message: "Label deleted successfully" });
+    }
   } catch (error) {
     console.error("Error deleting label:", error);
     res.status(500).json({ error: "Internal Server Error" });
